Events v2 progress
 - store guild as ulong instead of string
 - events v2 editor page
 - Table.Exists method for upgrade path

+++ FC.Manager.Server/Services/Eventsv2Service.cs
@@ -8,12 +8,12 @@ namespace FC.Manager.Server.Services
 	using System.Collections.Generic;
 	using System.Threading.Tasks;
 	using FC.Data;
-	using FC.Events;
+	using FC.Eventsv2;
 	using FC.Manager.Server.RPC;
 
-	public class EventsService : ServiceBase
+	public class EventsV2Service : ServiceBase
 	{
-		private Table<Event> eventsDb = new Table<Event>("KupoNuts_Events", 1);
+		private Table<Event> eventsDb = new Table<Event>("Events", 2);
 
 		public override async Task Initialize()
 		{
@@ -24,31 +24,38 @@ namespace FC.Manager.Server.Services
 		public async Task<List<Event>> GetEvents(ulong guildId)
 		{
 			Dictionary<string, object> search = new Dictionary<string, object>();
-			search.Add("ServerIdStr", guildId.ToString());
+			search.Add("GuildId", guildId);
 			return await this.eventsDb.LoadAll(search);
 		}
 
 		[GuildRpc]
 		public async Task DeleteEvent(ulong guildId, string eventId)
 		{
+			Event evt = await this.eventsDb.Load(eventId);
+			if (evt == null)
+				return;
+
+			if (evt.GuildId != guildId)
+				throw new Exception("Attempt to delete another guilds event");
+
 			await this.eventsDb.Delete(eventId);
 		}
 
 		[GuildRpc]
 		public async Task UpdateEvent(ulong guildId, Event evt)
 		{
-			evt.ServerIdStr = guildId.ToString();
+			evt.GuildId = guildId;
 			await this.eventsDb.Save(evt);
 		}
 
 		[GuildRpc]
-		public Task<EventsSettings> GetSettings(ulong guildId)
+		public Task<FC.Events.EventsSettings> GetSettings(ulong guildId)
 		{
-			return SettingsService.GetSettings<EventsSettings>(guildId);
+			return SettingsService.GetSettings<FC.Events.EventsSettings>(guildId);
 		}
 
 		[GuildRpc]
-		public Task SaveSettings(ulong guildId, EventsSettings settings)
+		public Task SaveSettings(ulong guildId, FC.Events.EventsSettings settings)
 		{
 			// Don't let clients change this!
 			settings.Guild = guildId;
